Bug 1664771 - [docs] Remove bogus 'Expires' header when uploading docs, r=firefox-source-docs-reviewers,championshuttler
As per:
https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3.html

This header doesn't actually control when an object gets deleted. Rather it is when the object
expires in the browser cache (i.e, when the expiry is hit, the browser will reload the object).

Remove this as it's only causing confusion.

Differential Revision: https://phabricator.services.mozilla.com/D90162

UltraBlame original commit: 367c0d24e6bdff8e9760697122a778708410b8a8