Bug 1663755 - Remove `inherit-from-parent-environment` implementation r=ahal
This was originally meant to allow `virtualenv`s to use packages from a parent Python environment without having to re-install them. This turned out to not pan out as we would have liked, so we're going another way to solve the same problem. Bug 1660351 walked back a bunch of this logic; this patch deletes the rest of it.

Differential Revision: https://phabricator.services.mozilla.com/D89492

UltraBlame original commit: 4627a7a4713dbef81fb169254d08fefc52cbb235