Bug 1665184 - Wait serviceworker.txt to be created to prevent test_service_workers_at_startup intermittent failures. r=asuth
The intermittent failure seems to be easily reproducible locally by running the marionette test multiple times in a row
using `mach marionette-test --run-until-failure ...`.

Explicitly waiting for the serviceworker.txt file to be created in the Firefox profile
before restarting the instance seems to be enough to allow the test to complete successfully
with --run-until-profile, and so the reason behind the intermittent failure is very likely
due to the Firefox instances quitting before the registered workers are being actually being
stored in the profile.

This patch does add a small change to the marionette test that makes sure that the test case
is going to restart the Firefox instance after the serviceworker.txt file has been created
in that Firefox profile.

Depends on D90189

Differential Revision: https://phabricator.services.mozilla.com/D90369

UltraBlame original commit: e5b242bfc7046fe3124a5624ac1a0f96787550c0