Bug 1665225 - Remove unnecessary to-main-thread dispatch r=alwu
The SMTC-thumbnail-update task is dispatched to main-thread from the
on-complete callback of the async image storing operation. However, the
callback is actually on the main-thread itself so the
thumbnail-update-task can be executed in that callback directly instead
of being dispatched. The task should be finished slightly faster if the
task dispatch is replaced by executing the task.

Differential Revision: https://phabricator.services.mozilla.com/D90331

UltraBlame original commit: 1a69e257dbd0aa45cf897993fe5b4eaa69a944e6