Bug 1656438 - Simulate audio input and verify audio output in MockCubeb. r=padenot
With this patch, AudioGenerator is used to create a sine tone audio input to a
duplex stream. In parallel, the AudioVerifier is used to verify that this sine
tone exists in the output (on demand).

This is the first approach. Fancier generators/verifiers can be future work.

Depends on D85554

Differential Revision: https://phabricator.services.mozilla.com/D85555

UltraBlame original commit: 64db6defa1538e890e9321a6756a84367d20c743