Bug 1665351 - Disable CFG for plugin hook stub calls r=handyman
As we saw in bug 1598119, 32-bit nop-space patches aren't compatible with clang 11's CFG because they return to the second instruction of the hooked function.

The `FunctionHook`s for plugins were pulling raw function pointers out of the interceptor stubs, so they didn't get the benefit of the stub's `operator()` that already has the CFG annotation.

As a bandaid, this patch marks all users of `BasicFunctionHook::OriginalFunction()` with the CFG annotation as well. A more thorough fix might be to somehow pass through to the stub's `operator()`, but we need something before merge day and I'm not confident in being able to do that regression-free in time.

Differential Revision: https://phabricator.services.mozilla.com/D90629

UltraBlame original commit: 9c8401590bab1719ff59b0bb4460048b5634b770