Bug 1665323. Make sure the visual viewport offset always takes on valid values. r=kats
The problems that it causes that are referred to in the patch is that we use the vvoffset and GetScrollRangeForUserInputEvents to compute the scrollbar cur, min, and max attributes. If we have a non-zero vvoffset when we are zooomed backed out, GetScrollRangeForUserInputEvents will be a zero range and so the cur attribute (non-zero)  will be larger than the max attribute (0).

This then causes nsSliderFrame::AttributeChanged to call ScrollByWhole at

https://searchfox.org/mozilla-central/rev/30e70f2fe80c97bfbfcd975e68538cefd7f58b2a/layout/xul/nsSliderFrame.cpp#204

Which we definitely do not want.

Differential Revision: https://phabricator.services.mozilla.com/D90372

UltraBlame original commit: 8081a13f8d251410bb9a76f513520333b9dbfd57