Bug 1637351: Add onKill handler to GVE; r=geckoview-reviewers,agi
Since GVE is supposed to be an exemplar for using GV the Right Way (TM), we
add an `onKill` handler that checks whether the killed session is current, and
does not restore the session if it is in the background.

When switching sessions, we add a check to see whether the session was closed,
and reopen it if so.

Differential Revision: https://phabricator.services.mozilla.com/D91465

UltraBlame original commit: b195f27085beaa5fc62296efa15933a1bf57b1b2