Bug 1665553 - Undefined value use in `moz_gtk_button_paint()`. r=stranskyredhat.com.
nsNativeThemeGTK::GetGtkWidgetAndState looks as if it should write to
out-parameters `*aState` and `*aWidgetFlags` any time it returns `true`.  But
it is possible to observe that `*aWidgetFlags` is uninitialised when `true` is
returned.  This simple patch zero-initialises `*aWidgetFlags` at the start, in
the same way that `*aState` is already zero-initialised at the start.

Differential Revision: https://phabricator.services.mozilla.com/D90697

UltraBlame original commit: 8f6e6ecd05961a0a0766243024470b7bb30963e8