Bug 1639153 - Part 6.0: Force Ion to preserve its tls. r=lth
Let's see following code and let's assume that wasm-compiler is ion:

call foo
call bar

Ion can clobber tls inside foo and then it can go with clobbered tls into bar.
There will be a crash if bar uses tls. At compile-time we don't know whether bar will use tls or not.

It works when we restore tls each time when we are returning from a function because of the current frame structure.
But now, when we want to get rid of Frame::tls we should guarantee that Ion doesn't clobber tls inside a function.
In fact we forces Ion to restore tls iff it clobbers it and it is a rare case.

Baseline doesn't need such logic because of its private stack slot for tls.

Differential Revision: https://phabricator.services.mozilla.com/D83061

UltraBlame original commit: 0e268b8d01fe869fc9c488e4dc627b5962a9ea43