Bug 1667084 - [devtools] Rewrite browser_aboutdebugging_addons_debug_popup.js to close toolbox after message evaluation r=daisuke
Depends on D92090

This intermittent was caused because we are closing the toolbox without waiting for the console to evaluate the message completely.
While we could a smaller change to simply close the toolbox later, this rewrite is intended to make the test easier to read.
Especially:
- it extracts some logic in dedicated helpers
- it removes complicated async interactions between the main test method and the testToolboxScript method

Differential Revision: https://phabricator.services.mozilla.com/D92092

UltraBlame original commit: d4f892fc52cf9d2644b1df679e850da635228ec9