Bug 1664804 - Compute whether top layer content is opaque. r=miko
The default styling for a ::backdrop pseudo element results in it being fully opaque and occluding all the rest of the page.
This allows us to detect that case early, and skip doing any work for the rest of the page.

Differential Revision: https://phabricator.services.mozilla.com/D91669

UltraBlame original commit: 237da1bcdd52e1a43eb6c74a5edf0429592f7f66