Bug 1669196 - Guard a test on word width.  r=jseward
A test introduced in bug 1669003 should only be run on 32-bit systems
for now because (a) that's what it's intended for and (b) it will fail
on aarch64 due to bug 1666747 (bad unaligned writes at the end of
memory):

On 32-bit ARM, the test case performs aligned writes and thus tests
what it's supposed to test.  On 64-bit ARM, the write will be
unaligned and hence fail due to the latter bug.

Differential Revision: https://phabricator.services.mozilla.com/D92544

UltraBlame original commit: 402686aadec108248fe5c83e5ed6496174889c03