Bug 140718 - Add a "Print selection only" entry to the context menu. r=Gijs,jwatt,fluent-reviewers
This allows to use the feature from the context menu and is very
straight-forward.

This has the benefit of working with fission (as print from the focused
frame) and so on.

Making a checkbox on the UI on top of this is doable, but it is a lot
more complex (requires a bunch of both front-end and platform work)
because undoing this implies restoring the DOM which "print selection"
removes.

I think the context menu entry is very useful regardless of whether we
add a checkbox to undo it, and other browsers (with smaller context
menus) seem to agree.

Differential Revision: https://phabricator.services.mozilla.com/D92205

UltraBlame original commit: 3fa0c4e1bb955a2ed81ba756df310a0b05665f98