Bug 1667728 - Add lookahead silence after resampling. r=bryce
Without this patch, the lookahead silence would be added to the source segment,
which is sampled at the input rate.

The lookahead is in the output rate.

Converting the lookahead to the input rate, and letting our regular encoding
logic convert it back to the output rate can lead to a rounding error for some
input rates. A rounding error here would lead to the last packet being too
short.

Differential Revision: https://phabricator.services.mozilla.com/D91725

UltraBlame original commit: bc3576399c26a34384b3e5e0351762bfc1943553