Bug 1669874 - Use `command -v` instead of `which` in mach. r=firefox-build-system-reviewers,andi,rstewart
The former is a POSIX shell construct, usually a builtin, and the latter
is an external program that is not guaranteed to be available.

When `which` is not available, a full ./mach build succeeds with this
change.

Differential Revision: https://phabricator.services.mozilla.com/D92864

UltraBlame original commit: 234ca9ad2c5798a54c8766b5c7bfc3e9cd90458d