Bug 1468250 - Block all versions of database.dll to stop the crash.  r=aklotz
We blocked the older versions of database.dll as Bug 1566109 in 2019,
but the same crash is still happening with the newer versions.
We decided to block all versions because crashing in the middle of printing
or file uploading is not acceptable.

Differential Revision: https://phabricator.services.mozilla.com/D92988

UltraBlame original commit: fc42ab7c51894dfef81cdc9ac4350864fb68c91e