Bug 1670452 - Make sure to reposition the popups if the preferred size changes, as that can further resize us. r=NeilDeakin
When the pref size changes, we set the bounds of the popup to the pref
size, which is the size of all the contents, usually. It's
SetPopupPosition what takes care of flipping / constraining / etc, so
make sure that happens.

Differential Revision: https://phabricator.services.mozilla.com/D93226

UltraBlame original commit: 6db4fe3478a614d8c355c495d99f61b70c9936c8