Bug 1670251 - Replace GCPtr with HeapPtr in data structures that can be destroyed outside of GC sweeping r=sfink
This means we don't need to use GCManagedDeletePolicy for these any more. I
don't think this will have a performance impact but it's possible.

It would be possible to add a TenuredHeapPtr wrapper for use in places where
you have a pointer kinds that are nursery allocatable but not used for nursery
pointers in practice. That could help mitigate any perf impact caused by the
scope changes.

Depends on D93162

Differential Revision: https://phabricator.services.mozilla.com/D93163

UltraBlame original commit: 9f69c283568885702338a93b078bbf4ad8ed2157