Bug 1660609 - [devtool] Fix browser_console_devtools_loader_exception.js failure on Beta. r=ladybenko.
This test was failing on Beta because we are using the targetList to wait for
a given source to be available, in order to avoid pending connection to the
server (that were happening when the targetList was trying to attach the thread
of the target), but on Beta, for the browser console, we don't listen for anything
with the targetList if the Browser Toolbox fission pref is disabled.
To fix the test, we check the pref and wait for the target only if it's enabled.

Differential Revision: https://phabricator.services.mozilla.com/D92877

UltraBlame original commit: 0ff3dd4be81f9cd06e4851733f4948bef1277eae