Bug 1670907 - moving buttons in customize mode should not disable them, r=mconley
When we move items in customize mode, we unwrap them, then move them, then wrap
them again. The unwrapping sets the `command` attribute back on the button, and
that also sets the `disabled` attribute on the button. The wrapping removes the
`command` attribute but leaves the `disabled` attribute.
When the `disabled` attribute is removed from the command, this does not
propagate to the button because the command attribute has not been put back.

To fix this, the patch avoids adding and immediately removing the command
attribute when moving items while in customize mode.

Differential Revision: https://phabricator.services.mozilla.com/D93338

UltraBlame original commit: bd5fbe4e35a37bbd7992fde127ab7d7254d3b1c4