Bug 1669956 - [devtools] Remove unused/unnecessary files from the reps folder. r=jdescottes.
There was an image folder, relic of an ancient time where
we needed to support the launchpad. We don't use those images anymore,
so the folder is removed.
The remaining files in the tset folder weren't used, so we remove
this folder as well.
Finally, there was a redux middleware folder, containing a thunk
and a waituntilService middleware. Those 2 modules already exist
in the codebase, so we can use those from the test that were using
the reps-specific version (with little adjustments).

This marks the end of the work for this bug.

Differential Revision: https://phabricator.services.mozilla.com/D93488

UltraBlame original commit: 5822afb8f110a05f06a9ff8472538b9bef28580f