Bug 1668545 - Advance lateWriteChecksStage to 3 outside Nightly r=mconley
This is a fairly trivial change. It's been on Nightly for quite a while
now. This will in practice turn on late write checking for early beta, as
the IOInterposer is not active past that point. We will then analyze any
late write reports and see if anything new turns up which we haven't
already seen and classified as harmless in Nightly.

Differential Revision: https://phabricator.services.mozilla.com/D93536

UltraBlame original commit: fdd37e7f4f5ba2b3def6581a26b67cdf09238872