Bug 1664719 - Pt 4 - Refactor complex stacking context clip detection. r=nical
Previously, we relied on walking the built clip-chain for a
stacking context to determine if it needed an intermediate surface
due to presence of a complex clip.

However, future patches in this bug need to delay clip-chain creation
for a primitive until the tile-cache slice for the primitive is
known. This change means we don't rely on the clip-chain early
during push_stacking_context (but should have no functional change).

Differential Revision: https://phabricator.services.mozilla.com/D93904

UltraBlame original commit: 329e0bb968c19734b21309d726ff79994ef09684