Bug 1365886 - [marionette] Allow performActions to operate on chrome elements r=marionette-reviewers,jdescottes,whimboo
Change error handling and initialization on the server side.

On the Marionette Python client side, add a `kind` attribute to HTMLElement
to distinguish chrome elements from content elements in the action sequence
sent to the server. This change is necessary for `performActions` in contrast to
other command implementations because of the extra parsing step done by
`actions.Chain.fromJson` on the server side.

Note that for the time being Marionette's ReferenceStore does not
distinguish chrome and content elements (Bug 1672788), so this client-side
change is correct but not strictly necessary.

Differential Revision: https://phabricator.services.mozilla.com/D93778

UltraBlame original commit: 3a4b62af959713ca7229195de50b281b17a57a2e