Bug 1672770 - Generate better x86 ion code for v128 load-and-widen. r=jseward
Load-and-widen can be expressed as a single instruction on x86, so
let's specialize this path (in Ion) to avoid temp registers and extra
instructions.

We do what we did for zero-extend and splat: add an attribute to the
MemoryAccessDesc and let the wasmLoad logic take care of the
instruction selection.

Differential Revision: https://phabricator.services.mozilla.com/D94551

UltraBlame original commit: 857eeed4ed64042877bbffe719cb3e847eb961fa