Bug 1673364 - fix test_dns_disabled.js r=necko-reviewers,dragana
The test was using network.dns.localDomains to check that we don't
call into the platform DNS resolver when the network.dns.disabled pref
was set - but since the localDomains pref rewrites hostnames to localhost
and we now don't call into GetAddrInfo for local domains, the test
failed.
I changed it so it uses the NativeDNSResolverOverride to register an IP for
foo.bar instead of relying on localDOmains.

Depends on D94726

Differential Revision: https://phabricator.services.mozilla.com/D94727

UltraBlame original commit: 6df4ac02675fbd97ff06aaba7b44cf0a0c9e3018