Bug 1672625 - Make monochrome support on GTK work like on Mac. r=jfkthame
We do it in nsDeviceContextSpecG since that's what actually consumes the
settings. On GTK the options need to be prefixed by cups- for some
reason in order to work, so factor out a macro listing the options and
do the NSString / cups- prefixing in the platform specific places.

Differential Revision: https://phabricator.services.mozilla.com/D94479

UltraBlame original commit: 4f179bf502cedd18bf16a625477d5902e4304c89