Bug 1673723 - Test marker schema from xpcshell tests; r=gerald,julienw
Moving markers definitions to where they are used means that it's harder
to unit test them from gtests. xpcshell tests have the benefits of exercising
the markers in a more realistic setting. However, we still need to test the
output of the schema. This patch adds a new capability to do a deep equality
check of the schema results. The utility function attempts to make the output
more readable by humans, and suggest how to fix errors. This will hopefully
make it easy for users to update and maintain their markers.

Differential Revision: https://phabricator.services.mozilla.com/D94890

UltraBlame original commit: c386129a2f645e354a64d154ccdecd252e2943e5