Bug 1674279 - When determining whether to start a fling, require the velocity to be greater than apz.fling_min_velocity_threshold (change >= to >). r=botond
This only really makes a difference when the pref is set to zero: Now we will no
longer attempt to kick off no-op flings. Many of the APZ gtests are manually
setting the pref to zero, and were unintentionally kicking off no-op flings.
As a result, they were crashing on an assertion in APZFlingPhysics which checks
that flings always start with a non-zero velocity.

When the pref is set to a non-zero value, which is the case for our users, it
doesn't really make a difference whether we check > or >=, because users are
very unlikely to hit the "==" case. So this patch shouldn't affect the user
experience.

As a result of this change, the following test starts passing on Android:

APZCFlingStopTester.FlingStopTap

There are a lot more tests that were hitting this assertion, but those are still
failing for other reasons.

On non-Android, this change causes the following test to fail:

APZCPinchGestureDetectorTester.Panning_TwoFingerFling_ZoomDisabled

That's because the regular velocity tracker determines a velocity of zero, and
the test checks that a fling is started. This test will be fixed in an upcoming
patch in this series.

Differential Revision: https://phabricator.services.mozilla.com/D95246

UltraBlame original commit: a503b1f051f974a89e705fccbdb56d07bdbc4ed3