Bug 1418001 - Generate analysis for file inclusions. r=asuth,andi
This introduces a new mangled symbol of the form FILE_<hash>. It's defined
on line 1 of each visited source file, and it's referenced in places that
the preprocessor tells us there was an #include. Both angle-bracket includes
and quote includes are supported here.

Differential Revision: https://phabricator.services.mozilla.com/D95529

UltraBlame original commit: a06787957417530385940b0216c240dbd5867f58