Bug 1674398 - Follow-up: Correct base path and pattern for comm mozharness.zip. r=ahal
The original change was tested as working on try/try-c-c, but was done during
the window when bug 1673992 had been backed out on mozilla-central. As a result,
the mozharness.zip that was tested had the "mozharness/" top level directory and
everything worked.

Differential Revision: https://phabricator.services.mozilla.com/D95741

UltraBlame original commit: fb42612e38105ad3ace7354b66ccb4315494bb65