Bug 1673018 [wpt PR 26260] - Remove Signature-based Resource Loading Restrictions, a=testonly
Automatic update from web-platform-tests
Remove Signature-based Resource Loading Restrictions

Signature-based Resource Loading Restrictions is an experimental
feature of Content Security Policies for allowing execution of
scripts only if they have a valid signature. The feature was
prototyped, but it looks like we are not shipping it as it is
for the time being.

Let's remove it from now, we will get it back in the future when
we decide to come back to this.

Bug: 753349
Change-Id: I8fb808999144c32d6fa94ae5bf9299e1da275e3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2494882
Commit-Queue: Antonio Sartori <antoniosartorichromium.org>
Reviewed-by: Mike West <mkwstchromium.org>
Cr-Commit-Position: refs/heads/master{#821222}

--

wpt-commits: 8fefedccb9ae49bddd9c84d22b21177f156a561e
wpt-pr: 26260

UltraBlame original commit: 6ca7db736acff063858888eb1548284244b85887