Bug 1675069 - Use nsDependentCSubstring rather than ptr+length in unic-langid-ffi. r=zbraniecki
Turns out we have `nsCStr` just for this already. I wrote this to
confirm it works as expected and it does, so...

The code doesn't deal with raw pointers anymore so we can remove the
helpers module and so on.

Differential Revision: https://phabricator.services.mozilla.com/D95866

UltraBlame original commit: 3e24b2c81d24f3a62cf4918687d52156f83c540b