Bug 1665389 - create bookmarks on the toolbar by default, r=mak,jaws
This adds a pref containing a parent GUID, a lazy pref getter that validates
that GUID (asynchronously), and starts using the pref from the
PlacesCommandHook.

It also sets the future default (toolbar) into firefox.js, and overrides
that on the default branch when the new 2020 bookmarks pref is not set.

Finally, it sets the pref to the unfiled default for existing profiles
with a migration. If we end up delaying shipping, we'll need to
update that migration - but I don't see a way around that.

Differential Revision: https://phabricator.services.mozilla.com/D94500

UltraBlame original commit: 09f049232773a7bfefa5822296547b51d083c16e