Bug 1675323 - Apply scrollbar-width to the content select scrollbox. r=mconley
(Instead of applying it directly to the menulist).

Since it's not inherited, we need to get the right value over to the scrollbox.
For that, export it from the arrowscrollbox, with the same name as the places
menupopup uses for the same purpose.

Differential Revision: https://phabricator.services.mozilla.com/D95887

UltraBlame original commit: 93b19729112e9770cf1cadd5454a1a1bed4a30ca