Bug 1673641 - [devtools] Fix browser_browser_toolbox_evaluation_context.js failure on beta simulation. r=ladybenko.
The test was failing on Beta because we weren't setting the devtools.webconsole.input.context
pref, which is true only on Nightly.
This patch flip this pref in the test, which seems to make the test pass.

Differential Revision: https://phabricator.services.mozilla.com/D96061

UltraBlame original commit: c59f6a33c7753b2cad2278a689df585a47ad40bb