Bug 1582860 - Convert bogus fatal assertions about missing Places ids into warnings. r=Standard8
These assertions were intended to protect us from trying to build a result
node from a non existing database id. In practice it is totally possible that
between the time an entry is added and the time it is notified, something else
executes a removal of the same entry.
The assertions are thus converted into warnings, so that we can still notice
them when running manual tests, but they won't cause test failures, since it's
a condition that can happen during test runs, but not a bug.

Differential Revision: https://phabricator.services.mozilla.com/D96224

UltraBlame original commit: f7bf357cf2bdbb3b398d4506874816b7c65d10e7