Bug 1675074 - Use constexpr for all well-known ParserAtoms. r=arai
This removes the LifoAlloc from WellKnownParserAtoms and instead using
constexpr to store the atom data in WellKnownParserAtoms_ROM so that this is
all evaluated at startup time and live in read-only memory.

The WellKnownParserAtoms itself still maintains fields and a hash-table that
are not in the ROM, but exist once per JSRuntime.

Differential Revision: https://phabricator.services.mozilla.com/D96254

UltraBlame original commit: e83e1fece2dd9f57a16a3e05056781d28cbe5c9d