Bug 1654817: Part 2 - Bump dom.ipc.processCount pref to 3 on Nightly; r=geckoview-reviewers,agi
I want to start with 3 so that we initially get a good look at how things run
with multiple processes, but when we easily hit the limit. We'll increase to
our final process count in a future bug.

Differential Revision: https://phabricator.services.mozilla.com/D94884

UltraBlame original commit: bf721cd8ab144c912351fe278cb5c580df3cc194