Bug 1676057 - Disable scrollbar buttons when non-native theme is in use on Linux. r=emilio,spohl
Whether scrollbar buttons are shown is currently determined by
LookAndFeel values, and these are exposed through system metric media
features.  LookAndFeel values are process-wide, but which theme is in
use is document-specific.  (It can be the platform's native theme or the
native basic theme, depending on Document::ShouldAvoidNativeTheme.)  So
if nsNativeBasicTheme on Linux needs to not support scrollbar buttons,
we can't do this by forcing the LookAndFeel values to be false.

So instead, we skip creating the native anonymous content for the
scrollbar buttons when the non-native theme is in use on Linux.  There's
no need to support dynamic changes, since whether the non-native theme
is in use for a given document never changes.

Differential Revision: https://phabricator.services.mozilla.com/D96345

UltraBlame original commit: 89bf767253601652e5f4721f7e6f8ba435a8cdbc