Bug 1673391 - Introduce new masm abstractions for binop with constant. r=jseward
This introduces the instruction encoders needed for the next patch,
which will emit specialized code for operate-with-constant for many
SIMD instructions.

There's possibly a discussion to be had here about whether it is
necessary to expose these abstractions in MacroAssembler.h, since they
are x86-specific.  I decided it was good hygiene to do so, as (a) the
baseline compiler could in principle make use of them and (b) it's not
out of the realm of the possible that we'll want Ion SIMD support on
other platforms in the future.  I could be convinced otherwise.

Differential Revision: https://phabricator.services.mozilla.com/D95505

UltraBlame original commit: b1594250df3328bc439b2f40dc68a91c93eea7bb