Bug 1670328 - remove crossbeam usage from SwCompositor. r=mattwoodrow
We seem to be spending a significant amount of time inside crossbeam channels
sending jobs across. Sending multiple bands for a given job tends to relock the
channel multiple times when ideally we only want to queue the job once, or lock
the channel once (neither of which we can conveniently do with crossbeam). To
alleviate this, I've implemented a more custom solution with a mutexed VecDeque
and some Condvars.

Differential Revision: https://phabricator.services.mozilla.com/D96520

UltraBlame original commit: e111e9ce37542e1071f442af24631e03c19f92eb