Bug 1675205 - Report telemetry when load an empty file via a JAR archive r=zbraniecki
This will report when a telemetry even whenever we load an empty file via a
JAR channel. It will not catch situations where onStopRequest is called with
an error code, or when the listener of the channel does not consume the stream.

Differential Revision: https://phabricator.services.mozilla.com/D96412

UltraBlame original commit: b5cf0a593e64a88d1d8766b3a1e38f94bdb32394