Bug 1672072 - P2. Remove unused DecoderDoctorDiagnostic from CreateDecoderParam. r=mattwoodrow,bryce,padenot
The DecoderDoctorDiagnostic is only ever used with PDM::Supports; when creating a decoder this object would always be nullptr.
Attempting to set it would have resulted in UAF as a DiagnosticDecoderDecoder is typically an object allocated on the stack and most of the code creating the decoders is async.

Differential Revision: https://phabricator.services.mozilla.com/D96358

UltraBlame original commit: fbf0b51176552b077a35b38bf735a08ad9983832