Bug 1669724 - Fix the mistake in `runQueryContentEventRelativeToInsertionPoint` in `window_composition_text_querycontent.xhtml` r=m_kato
This is simple mistake at writing the test.

And also, once fixing the mistake, another bug appears.  That is, the second
composition string is not "b", but the following code assumes so.  Therefore,
this fixes it too.

Depends on D97269

Differential Revision: https://phabricator.services.mozilla.com/D97270

UltraBlame original commit: 8a95095384a591fa0b5e18d7bf5960b3b3e07985