Bug 1677520 - Add a FramesVerifiedEvent to MockCubeb and re-purpose the FramesProcessedEvent. r=padenot
Prior to this patch the FramesProcessedEvent only exposed the number of frames
that had been verified non-silent. This meant there had to be a verifier and an
audio output hooked up to the audio input fed by the generator, for the event to
report any frames at all.

With this patch the FramesProcessedEvent reports the number of frames processed
by the fake audio thread, and the FramesVerifiedEvent the number of non-silent
frames seen by the verifier.

Differential Revision: https://phabricator.services.mozilla.com/D97307

UltraBlame original commit: e820d352df4cb1cb1ebaed4ca6d3cb95f3c886d8