Bug 1669990 - SubDialog: Use system event listener for escape key. r=Gijs
If a <select> in a SubDialog closes its popup when the user hits the esc key, it calls `preventDefault()` on the event. The SubDialog code didn't see this preventDefault, because it was using a normal event listener, while the select code uses a system event listener.

https://searchfox.org/mozilla-central/rev/c2e3ac11be4837718c2604e58085fbc8252b69dd/layout/forms/nsListControlFrame.cpp#925,2058,2068-2069,2075-2076

Differential Revision: https://phabricator.services.mozilla.com/D93253

UltraBlame original commit: 5066eae18c2da8adecfa968a040b89af7acb138c