Bug 1673716 - Ensure the about:home startup cache can only be used by one BrowsingContext. r=Gijs
It is theoretically possible for two BrowsingContexts to be loading about:home
at the same time during startup (for example, if the user passed in a series
of URLs to open via the command-line, including multiple about:home's).

This patch ensures that only one of those BrowsingContexts (the first to load)
gets to consume the cached streams.

Differential Revision: https://phabricator.services.mozilla.com/D97518

UltraBlame original commit: dc9ac32cc19bc60cc59ea71f375660555dda81da