Bug 1676559 - Pt 5 - Refactor `RenderTask` to move creation logic to RenderTaskKind r=nical
With follow up patches, we want to retain the existing logic and
structures in RenderTaskKind, but be able to port and use them in
a new render task graph structure. This is simpler if we move as
much logic out of `RenderTask` as possible.

- Move simple RenderTask create methods into RenderTaskKind
  (complex ones will be done as a separate follow up)
- Move `write_task_data` and `write_gpu_blocks` into RenderTaskKind
- Move some static helper methods (e.g. to `BlurTask`)

Differential Revision: https://phabricator.services.mozilla.com/D97527

UltraBlame original commit: 46da71790703187e099950bd23a7563bde6693e2