Bug 1679700 - Cast int parameter in get_yuv_offset_vector to float, to work around a shader compilation bug with integer equality comparisons on macOS Intel. r=jrmuizel
We've encountered similar bugs before, e.g.
https://github.com/servo/webrender/wiki/Driver-issues#3540---mac-glsl-compiler-bug-with-integer-comparison-take-two

We don't want to change this `if` to a `switch` because glsl-opt compiles it
down to an if statement anyway, just a more complicated one (which was triggering
other bugs on other platforms, see bug 1678924).

Differential Revision: https://phabricator.services.mozilla.com/D98208

UltraBlame original commit: fdf86dfacb7ca6116997bf925d748a7d347853e7