Bug 1678636 - [devtools] Prevent paused overlay to be shown after resuming. r=ladybenko.
Before showing the pause overlay, we wait for the page to be fully loaded.
This introduced some delay, and at the moment we were showing the overlay, the
debugger could not be paused anymore.
This patch adds another check to ensure the debugger is paused, as well as a
test to avoid regression."

Differential Revision: https://phabricator.services.mozilla.com/D97778

UltraBlame original commit: f29982da344032067e461641e580c39298190ff4