Bug 1679873 - Pass closures to IOUtils::RunOnBackgroundThread r=nika
Passing in a callable and its arguments to RunOnBackgroundThread was fine
before we started passing in already_AddRefed<nsIFile>. However, if we fail to
get an event target, then we drop all our arguments and in debug builds
already_AddRefed<T> asserts that it was moved out of.

Now we require callers of RunOnBackgroundThread to pass in a closure that
contains all the bindings they require, which has the added bonus of making
lifetimes more explicit. As part of this refactor, all the `ReadSync` etc
methods now take raw nsIFile* pointers since they are called in a scope that
has them refcounted.

Differential Revision: https://phabricator.services.mozilla.com/D98273

UltraBlame original commit: 6c71b9a049a323c1db4ffbeba70ec39b6b57af9b