Bug 1680274 - enabling the experiment pref during first startup should show the import button in the toolbar, r=jaws
This allows the pref to cause us to add the button later during the uptime of the first run. Note that this
would still require us to apply the experimental pref during the first run - it won't work if the experiment
only applies on the next run - the only way to do that would be to save even more state in a pref or similar,
and that didn't seem worth it.

Differential Revision: https://phabricator.services.mozilla.com/D98606

UltraBlame original commit: 397a9ae6efd49267f9dd72196d51d50b26e991ee