Bug 1673553 part 85 - Remove JSOp::IterNext. r=iain
This was a hint for IonBuilder to prevent loop restarts. Although this still lets
us do an infallible unbox in Warp, there's not much of a perf improvement on a for-in
micro-benchmark so it's probably fine to remove this now.

Differential Revision: https://phabricator.services.mozilla.com/D98589

UltraBlame original commit: 30953134ae145a073d80e510119556f40513ab8a