Bug 1678542 - Placeholder for patchable label on ARM64 must be ImmWord, not ImmPtr. r=nbp
This is not a functional change but allows switches to be emitted for wasm:
wasm disallows the use of an ImmPtr for movePatchablePtr but allows ImmWord,
this is in order to guard against embedded pointers in the serializable
machine code.  For a patchable word that will always be patched after
code is deserialized, we can use ImmWord, as we do on other platforms.

Differential Revision: https://phabricator.services.mozilla.com/D97771

UltraBlame original commit: f49a0516ef053a85151b439a15f2046d7cf54d36