Bug 1681448 - [devtools] Fix evaluation result message ordering. r=bomsy.
When evaluating an expression containing console api calls,
it might happen that the result message is handled by the
client before the console api messages.
This is due to the fact that those messages are handled by
the resource watcher, which throttles resources, whereas
the result message is directly added from the webconsole
frontend.
In the console reducer we already have code handling out-of-order
messages, but it is only triggered for messages with different
timestamps; in our case, the timestamp of the result message
might be the same as the one of the console api messages.

To fix this issue, we simply add an extra millisecond on the
evaluation result packet, which will ensure it will have a
distinct timestamp from the console api messages it might
have triggered, with minimal risk of side effects.

The existing test for ordering is enhanced by logging multiple
messages, and it is failing quite frequently without the fix.

Differential Revision: https://phabricator.services.mozilla.com/D99205

UltraBlame original commit: 1b2cb64bd1ce667cd57c33524720ee471847e9fc