Bug 1673553 part 100 - Move kind field from ICStub to ICFallbackStub. r=iain
If a stub has a non-null next_, it's a CacheIR stub. Else it's a fallback stub.
This means the Kind is only relevant for fallback stubs and we can move it there,
removing a word from each CacheIR stub.

The use of CacheIR_Regular here is a bit awkward, a later patch will rename this to
ICCacheIRStub to make this more natural.

This also removes the extra_ field that's no longer used.

Differential Revision: https://phabricator.services.mozilla.com/D99341

UltraBlame original commit: 516536a2d9a23adbdb536e3834b2adfa426bcdf1