Bug 1680440 - [devtools] Remove backward compatibility comments for descriptor watcher traits. r=jdescottes.
The watcher trait might not be set on other descriptors, so
we want to keep them until we can get rid of the generic client
code that is handling them. This can probably be done as part of Bug 1680280.

Differential Revision: https://phabricator.services.mozilla.com/D98737

UltraBlame original commit: 4fb557f04183c0a792c4fb99fa3ab16f1b141c65