Bug 1680769 - Fix incorrect returned size in some cases. r=jnicol
When deallocating the last item of the region the region's size is changed in free and we end up returning the wrong value.
The faulty code is actually removed in a later patch from this series but I'd like to fix the regressing now and land the patch that removes the bad code during the next train.

Differential Revision: https://phabricator.services.mozilla.com/D99458

UltraBlame original commit: 11ffdf919a4a1d0a29a7718a3a53a8a4b4468434