Bug 1681806: Add test r=jandem
We did not have test coverage for the code in `InvalidationBailout` handling a failure in `BailoutIonToBaseline`.

The testcase produced by the fuzzer took over a minute to run on my machine. I managed to get it down to about 10 seconds in debug.

Differential Revision: https://phabricator.services.mozilla.com/D99410

UltraBlame original commit: 10bfaf460d375fde0860f0ac7391530c20d82641