Bug 1678357 - part1 : activate the controller when it starts playing. r=chunmin
We currently activate the controller when it first becomes audible, which is possible to drop the first `playbackstatechange` event if the task to dispatch that evnt runs before we activate the controller.

The audible change always comes later than the playback change, so if we want to 100% ensure the first `playbackstatechange` can happen after `activated` event, then we have to activate the controller earlier. Therefore, this patch would activate the controller once it becomes playing.

Note that, even if we remove the audible check in media controller, we would still only control audible media because we have filtered out inaudible media in the content process.

Differential Revision: https://phabricator.services.mozilla.com/D98829

UltraBlame original commit: 903b57c15ff7361e309f38fd9dcf49d8e1bb0241