Bug 1675602 - Use WebAssembly's type system for TypedObject. r=lth
This commit removes TypedObject's type system (StructTypeDescr,
ArrayTypeDescr, ScalarTypeDescr) and replaces it with a
wasm::TypeHandle to a wasm::TypeDef inside the global
wasm::TypeContext. TypeDescr is left as a JSObject that
wraps a wasm::TypeHandle for a TypedObject. In the future
it will likely be re-used for a RttValue.

When instantiating a module, types are transferred to
the global context and TypeDescr objects wrapping the type handles
are created. The TypeDescr objects are placed in TlsData
for the TypeIdDesc corresponding to the StructType.

Differential Revision: https://phabricator.services.mozilla.com/D96223

UltraBlame original commit: e21edeca8575fc2eb0e42e2f92c51611d18e5759